Move common helpers into util, set bot status when clearing drinks
