+++ src/utils/constants.ts
@@ -1,0 +1,8 @@
+export const commands = `
+How to use Drunkcord! \n
+\`!cheers <drink_name>\` will add a drink\n
+\`!beers <beer_name>\` will add a beer with rich data\n
+\`!drinks\` will show how many total drinks have been drank\n
+\`!drunk\` will show who's drunk\n
+\`!closingtime\` will reset the drinks
+`;
